Chore: added the definitive migration
