+++ doc/v106_to_v110.md
@@ -2,6 +2,15 @@
 
 If you're using version v1.0.6 or minor, we recommend fixing all the deprecations before migrating to v1.1.0 for a smoother migration.
 
+## Parser
+
+The `RichTextParser` has been deprecated. We strongly recommend use DocumentParser instead.
+
+```diff
+- RichTextParser().parseDelta(delta);
++ DocumentParser().parseDelta(delta);
+```
+
 ## Paragraph
 
 The `Paragraph` now use another type of format, where the `Line` represents a new section of the `Paragraph`.
@@ -26,68 +35,134 @@ Paragraph(
 ),
 ```
 
-#### Before implementation
+## Generating a Document
 
-When we generates a `Document` using `RichTextParser`, this should be the result:
+Assume that you have this `Delta` that you want to parse:
 
 ```dart
-final doc = Document(
+final Delta delta = Delta()
+  ..insert('This is ')
+  ..insert('bold', {'bold': true})
+  ..insert(' and ')
+  ..insert('italic', {'italic': true})
+  ..insert(' text with ')
+  ..insert('custom color', {'color': '#FF0000'})
+  ..insert('\n\n', {'header': 1})
+  ..insert('This is a list item')
+  ..insert('\n', {'list': 'ordered'})
+  ..insert('Another list item')
+  ..insert('\n', {'list': 'ordered'})
+  ..insert('Third list item')
+  ..insert('\n')
+  ..insert('This is a ')
+  ..insert('link', {'link': 'https://example.com'})
+  ..insert(' to a website')
+  ..insert('\n');
+```
+
+When we generates a `Document` using `DocumentParser`, this should be the result:
+
+#### Old implementation will generate something like: 
+
+```dart
+Document(paragraphs: [
+  Paragraph(
+    lines: [
+      Line(data: 'This is '),
+      Line(data: 'bold', attributes: {'bold': true}),
+      Line(data: ' and '),
+      Line(data: 'italic', attributes: {'italic': true}),
+      Line(data: ' text with '),
+      Line(data: 'custom color', attributes: {'color': '#FF0000'}),
+    ],
+    blockAttributes: {"header": 1},
+    type: ParagraphType.block,
+  ),
+  Paragraph(
+    lines: [Line(data: '\n')],
+    blockAttributes: {"header": 1},
+    type: ParagraphType.block,
+  ),
+  Paragraph(
+    lines: [
+      Line(data: 'This is a list item'),
+    ],
+    blockAttributes: {'list': 'ordered'},
+    type: ParagraphType.block,
+  ),
+  Paragraph(
+    lines: [
+      Line(data: 'Another list item'),
+    ],
+    blockAttributes: {'list': 'ordered'},
+    type: ParagraphType.block,
+  ),
+  Paragraph(
+    lines: [
+      Line(data: 'Third list item'),
+    ],
+    type: ParagraphType.inline,
+  ),
+  Paragraph(
+    lines: [
+      Line(data: 'This is a '),
+      Line(data: 'link', attributes: {'link': 'https://example.com'}),
+      Line(data: ' to a website'),
+    ],
+    type: ParagraphType.inline,
+  ),
+  Paragraph.newLine(),
+]);
+```
+
+#### New standard implementation will generate something like: 
+
+```dart
+Document(
   paragraphs: [
     Paragraph(
       lines: [
-        Line(data: 'This is a segment '),
-        Line(data: 'of the text '),
-        Line(data: 'where we can see '),
-        Line(data: 'how works Paragraphs'),
+        Line(
+          fragments: [
+            TextFragment(data: 'This is '),
+            TextFragment(data: 'bold', attributes: {'bold': true}),
+            TextFragment(data: ' and '),
+            TextFragment(data: 'italic', attributes: {'italic': true}),
+            TextFragment(data: ' text with '),
+            TextFragment(data: 'custom color', attributes: {'color': '#FF0000'}),
+          ],
+        ),
+      ],
+      blockAttributes: {"header": 1},
+      type: ParagraphType.block,
+    ),
+    Paragraph(
+      lines: <Line>[Line.newLine()],
+      blockAttributes: {"header": 1},
+      type: ParagraphType.lineBreak,
+    ),
+    Paragraph(
+      lines: [
+        Line.fromData(data: 'This is a list item'),
+        Line.fromData(data: 'Another list item'),
+      ],
+      blockAttributes: {'list': 'ordered'},
+      type: ParagraphType.block,
+    ),
+    Paragraph(
+      lines: [
+        Line.fromData(data: 'Third list item'),
+        Line(
+          fragments: [
+            TextFragment(data: 'This is a '),
+            TextFragment(data: 'link', attributes: {'link': 'https://example.com'}),
+            TextFragment(data: ' to a website'),
+          ],
+        ),
       ],
       type: ParagraphType.inline,
     ),
-    Paragraph(
-      lines: [
-        Line(data: 'This is a header'),
-      ],
-      blockAttributes: {'header': 1},
-      type: ParagraphType.block,
-    ),
-    Paragraph(
-      lines: [
-        Line(data: 'This is another header'),
-      ],
-      blockAttributes: {'header': 1},
-      type: ParagraphType.block,
-    ),
+    Paragraph.newLine(),
   ],
 );
 ```
-
-#### The new standard implementation
-
-With the new standard, when we generates a `Document` using `RichTextParser`, this should be the result:
-
-```dart
-final doc = Document(
-  paragraphs: [
-    Paragraph(
-      lines: [
-        Line(fragments: [
-          TextFragment(data: 'This is a segment '),
-          TextFragment(data: 'of the text '),
-          TextFragment(data: 'where we can see '),
-          TextFragment(data: 'how works Paragraphs'),
-        ]),
-      ],
-      type: ParagraphType.inline,
-    ),
-    Paragraph(
-      lines: [
-        Line(data: 'This is a header'),
-        Line(data: 'This is another header'),
-      ],
-      blockAttributes: {'header': 1},
-      type: ParagraphType.block,
-    ),
-  ],
-);
-```
-
-## Line
