+++ lib/flutter_quill_delta_easy_parser.dart
@@ -4,6 +4,7 @@ export 'package:flutter_quill_delta_easy_parser/blocks/document.dart';
 export 'package:flutter_quill_delta_easy_parser/blocks/line.dart';
 export 'package:flutter_quill_delta_easy_parser/blocks/paragraph.dart';
 export 'package:flutter_quill_delta_easy_parser/blocks/text_fragment.dart';
+export 'package:flutter_quill_delta_easy_parser/easy_parser/document_parser.dart';
 export 'package:flutter_quill_delta_easy_parser/easy_parser/easy_parser.dart';
 export 'package:flutter_quill_delta_easy_parser/easy_parser/merger_builder.dart';
 export 'package:flutter_quill_delta_easy_parser/enums/enums.dart';
